+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support extending schema using partial GraphQL schema files.
 
+### Added
+- Handle `Illuminate\Auth\Access\AuthorizationException` exceptions.
+
 ## [10.0.0] - 2023-03-02
 
 ### Changed